requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.22
python-jose[cryptography]
bcrypt==5.0.0
openpyxl==3.1.5
